fix(disk/aix): use statfs for UsageWithContext on AIX
The AIX UsageWithContext implementations (cgo and nocgo) enumerated
mountpoints via perfstat or `df -v` and matched the requested path exactly,
so disk.Usage failed with "mountpoint %s not found" for any path that
wasn't a mountpoint (e.g. a subdirectory).

Add aix to the build tag of disk_unix.go so AIX reuses the existing
statfs(2)-based UsageWithContext, which works on arbitrary paths. AIX's
Statfs_t has all the block/inode fields needed. Move the FSType map and
getFsType helper to the shared disk_aix.go so disk_unix.go can resolve
getFsType on AIX.

Only AIX changes; other OSes are unaffected.
